+++ Python/stdlib_module_names.h
@@ -63,9 +63,8 @@ static const char* _Py_stdlib_module_names[] = {
 "_random",
 "_scproxy",
 "_sha1",
-"_sha256",
+"_sha2",
 "_sha3",
-"_sha512",
 "_signal",
 "_sitebuiltins",
 "_socket",
